Add downloadClientMods gulp task
[skip ci]
